Generalised LoadReflTBL Algorithm
The Algorithm will now allow for loading of old TBL files from REFL interfaces
as well as new TBL files which are formatted to have column headings on the first line.
Tests have been updated to reflect these changes however a few more test cases will need
to be added for the new TBL format.

Refs #15864
